http://issues.apache.org/activemq/browse/AMQCPP-30
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@508988 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/connector/openwire/commands/MessageAck.cpp
@@ -139,7 +139,7 @@ std::string MessageAck::toString() const {
     } else {
         stream << "   Object is NULL" << std::endl;
     }
-    stream << " Value of AckType = " << this->getAckType() << std::endl;
+    stream << " Value of AckType = " << (int)this->getAckType() << std::endl;
     stream << " Value of FirstMessageId is Below:" << std::endl;
     if( this->getFirstMessageId() != NULL ) {
         stream << this->getFirstMessageId()->toString() << std::endl;
